+++ package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@hookform/resolvers": "3.3.4",
+    "@paypal/react-paypal-js": "8.2.0",
     "@prisma/client": "5.9.1",
     "bcryptjs": "2.4.3",
     "clsx": "2.1.0",
