+++ src/components/index.ts
@@ -15,4 +15,6 @@ export * from './product/stock-label/StockLabel'
 export * from './ui/spinner/Spinner'
 export * from './ui/alert/Alert'
 export * from './ui/toast-notification/ToastNotification'
-export * from './provider/Provider'
+export * from './providers/Providers'
+export * from './orders/PaymentStatusAlert'
+export * from './paypal/PayPalButton'
