+++ src/components/paypal/PayPalButton.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import { PayPalButtons } from '@paypal/react-paypal-js'
+
+export const PayPalButton = () => {
+  return <PayPalButtons />
+}
